Pull up PushDownAndCombineOrderBy
